Add temporary DB setup scripts for end-to-end testing.
Add tests/db/setup-pgsql.sh and tests/db/setup-mysql.sh to spin up
temporary PostgreSQL and MariaDB instances using Unix sockets only.
Fix eventid column type mismatch in spamikaze-mysql-3.sql (SMALLINT
to INT) to match the eventtypes.id foreign key target.

+++ schemas/spamikaze-mysql-3.sql
@@ -45,7 +45,7 @@ INSERT INTO eventtypes (id, eventtext) VALUES
 CREATE TABLE ipevents (
   ip VARCHAR(45) NOT NULL,
   eventtime DATETIME NOT NULL,
-  eventid SMALLINT NOT NULL,
+  eventid INT NOT NULL,
   KEY idx_ipevents_ip (ip),
   KEY idx_ipevents_eventtime (eventtime),
   KEY idx_ipevents_eventid (eventid),
